+++ examples/17_proxy.rb
