+++ examples/all.rb
@@ -58,4 +58,4 @@ require_relative "15_conversations_features_comprehensive"
 require_relative "16_schema_generation"
 
 # Proxy
-require_relative "15_proxy"
+require_relative "17_proxy"
